initial try to redesign codebase
and improve README
and greatly improve docker image
and remove dead code
