+++ .gitignore
@@ -1,24 +1,3 @@
-# Rust
-
-/target/
-/*/target/
-
-# IntelliJ
-
-**/.idea/
-**/*.iml
-
-# VSCode
-
-/.vscode/*
-!/.vscode/launch.json
-*.code-workspace
-
-# Miscellaneous
-*.log
-*.ron
+/target
 .airshipper-env
 airshipper.db
-public/
-profiles/
-.packages/
