+++ README.md
@@ -1,24 +1,29 @@
 # Airshipper
+
+[![GitHub release)](https://img.shields.io/github/v/release/songtronix/airshipper?include_prereleases)](https://github.com/Songtronix/Airshipper/releases) [![License](https://img.shields.io/github/license/songtronix/airshipper?color=blue)](https://github.com/Songtronix/Airshipper/blob/master/LICENSE) [![Discord](https://img.shields.io/discord/449602562165833758?label=discord)](https://discord.gg/3mhSHMX) [![AUR version](https://img.shields.io/aur/version/airshipper?label=AUR)](https://aur.archlinux.org/packages/airshipper/) [![Custom badge](https://img.shields.io/endpoint?color=orange&label=Support%20me&url=https%3A%2F%2Fmoshef9.wixsite.com%2Fpatreon-badge%2F_functions%2Fbadge%2F%3Fusername%3Dsongtronix)](https://www.patreon.com/songtronix)
+
 A cross-platform Veloren launcher.
 
 ![Airshipper](https://www.songtronix.com/airshipper.gif)
 
 ## Features
 
-- [X] Update/Download and start nightly.
-- [X] Fancy UI with batteries included.
-- [X] Updates itself on windows.
+- [x] Update/Download and start nightly.
+- [x] Fancy UI with batteries included.
+- [x] Updates itself on windows.
 
 ## Download
+
 For all download options visit my website: [www.songtronix.com](https://www.songtronix.com)
 
 **NOTE:** Airshipper cannot be considered stable yet.
 
 #### Compile from source
+
 ```bash
 git clone https://github.com/Songtronix/Airshipper.git
-cd airshipper
-cargo run --bin airshipper --release
+cd Airshipper
+cargo run --release
 ```
 
 Make sure to have [rustup](https://rustup.rs/) installed to compile rust code.
