+++ client/wix/Dialog.bmp
