+++ client/wix/Veloren.ico
