+++ client/wix/main.wxs
