+++ publish_server.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+sudo docker build . -t docker.pkg.github.com/songtronix/airshipper/airshipper:latest
+sudo docker push docker.pkg.github.com/songtronix/airshipper/airshipper:latest
