+++ server/.airshipper-env.example
@@ -1,5 +1,5 @@
 AIRSHIPPER_GITLAB_SECRET=banana
-AIRSHPPER_ARTIFACT_STAGE=post
+AIRSHIPPER_ARTIFACT_STAGE=post
 AIRSHIPPER_TARGET_BRANCH=master
 AIRSHIPPER_TARGET_EXECUTABLES=linux,windows
 AIRSHIPPER_BUCKET_NAME=veloren
