+++ server/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "server"
+name = "airshipper_server"
 version = "0.2.0"
 authors = ["Songtronix <contact@songtronix.com>"]
 edition = "2018"
@@ -16,11 +16,10 @@ rocket_contrib = { git = "https://github.com/SergioBenitez/Rocket", branch = "as
 # Networking
 reqwest = "0.10.4"
 # Serialising/Deserialising
-serde = "1.0.101"
+serde = { version = "1.0.101", features = ["derive"] }
 serde_json = "1.0.41"
 # Database
 diesel = { version = "1.4.4", features = ["sqlite", "chrono"] }
-libsqlite3-sys = { version = ">=0.8.0, <0.13.0", features = ["bundled"] }
 rust-s3 = { version = "0.19.0" }
 # Error handling
 thiserror = "1.0.13"
@@ -31,3 +30,6 @@ derive_more = "0.15.0"
 regex = "1.3.1"
 lazy_static = "1.4.0"
 dotenv = "0.15.0"
+# bundle all the things!
+libsqlite3-sys = { version = "*", features = ["bundled"] }
+openssl-sys = { version = "*", features = ["vendored"] }
