+++ server/src/error.rs
@@ -11,11 +11,6 @@ use thiserror::Error;
 #[derive(Error, Debug)]
 pub enum ServerError {
     // Web facing
-    #[error("Invalid platform. Currently supported are windows and linux.")]
-    InvalidPlatform,
-    #[error("Invalid channel. Currently supported is nightly with upcoming support for releases.")]
-    InvalidChannel,
-    // Not really a serious error (see routes/api.rs)
     #[error("Respond with Status: {0}")]
     Status(Status),
 
@@ -39,22 +34,8 @@ impl<'r> Responder<'r> for ServerError {
 
         match self {
             // Web facing errors
-            ServerError::InvalidPlatform => {
-                resp.status(Status::BadRequest);
-                resp.sized_body(Cursor::new(format!(
-                    "Invalid platform. Currently supported are windows and linux."
-                )))
-                .await; // TODO: Do not hardcode (use enum_iterator or such)
-            },
-            ServerError::InvalidChannel => {
-                resp.status(Status::BadRequest);
-                resp.sized_body(Cursor::new(format!(
-                    "Invalid channel. Currently supported is nightly with upcoming support for releases."
-                )))
-                .await; // TODO: Do not hardcode (use enum_iterator or such)
-            },
             ServerError::Status(status) => {
-                resp.status(status).finalize();
+                resp.status(status);
             },
 
             // Internal errors
